Move convert to rainrate function into precipitation.py
